% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-package.R
\docType{package}
\name{abs-package}
\alias{abs}
\alias{abs-package}
\title{abs: Read csv tables exported from ABS TableBuilder and ABS Census TableBuilder.}
\description{
Read csv tables exported from ABS TableBuilder and ABS Census TableBuilder..
}
\author{
\strong{Maintainer}: Amarin Siripanich \email{email@email.com}

}
\keyword{internal}
