% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablebuilder.R
\name{abs_read_tb}
\alias{abs_read_tb}
\title{Read tables exported from ABS Table Builder}
\usage{
abs_read_tb(x, exclude_total = TRUE)
}
\arguments{
\item{x}{path to a csv file exported from ABS Table Builder.}

\item{exclude_total}{exclude rows with Total}
}
\value{
a data.table
}
\description{
This provides an simple function to read in tables from Table Builder without
having to 'tidy' the tables first. The basic idea is that the function looks
for the first row that has its cell value equal to 'Count' which is what appeared
in all Table Builder tables. The 'Count' cell usually appear in as the last
column on the table and never the first.
}
\examples{

data_dir <- system.file("extdata", package = "abs")
test_csv <- file.path(data_dir, "tb1.csv")
mytable <- abs_read_tb(test_csv)

}
