% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb-tools.R
\name{tb_distribute_not_stated}
\alias{tb_distribute_not_stated}
\title{Reweight rows with not stated}
\usage{
tb_distribute_not_stated(x, zone_col)
}
\arguments{
\item{x}{a data.frame}

\item{zone_col}{a character indicates the zone column for groupping}
}
\value{
a data.frame
}
\description{
Reweight rows with not stated
}
