% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb-tools.R
\name{tb_range}
\alias{tb_range}
\title{Create a range from a numeric vector of size two}
\usage{
tb_range(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{

}
\description{
If the length of the vector is more than two then the vector gets returned.
}
