% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb-tools.R
\name{tb_remove_overseas_visitors}
\alias{tb_remove_overseas_visitors}
\title{remove overseas visitors}
\usage{
tb_remove_overseas_visitors(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
remove overseas visitors
}
