% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb-tools.R
\name{tb_remove_rows_where}
\alias{tb_remove_rows_where}
\title{remove rows where 'var' in 'value'}
\usage{
tb_remove_rows_where(x, var, value)
}
\arguments{
\item{x}{a data.frame}

\item{var}{column name in the data.frame to be evaluated}

\item{value}{a value or a vector of values that should be removed}
}
\value{

}
\description{
remove rows where 'var' in 'value'
}
