% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb-tools.R
\name{tb_remove_rows_with}
\alias{tb_remove_rows_with}
\title{remove rows with 'value'}
\usage{
tb_remove_rows_with(x, value)
}
\arguments{
\item{x}{data.frame}

\item{value}{a vector}
}
\value{
data.frame
}
\description{
remove rows with 'value'
}
