% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb-tools.R
\name{tb_simplify_names}
\alias{tb_simplify_names}
\title{simplify table builder headers}
\usage{
tb_simplify_names(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
simplify table builder headers
}
